+++ blockfrost-api/CHANGELOG.md
@@ -22,6 +22,11 @@
   * `ScriptType`
     * `Plutus` type is now `PlutusV1`
     * adds `PlutusV2`
+  * `ProtocolParameters`
+    * `coinsPerUtxoWord` is now deprecated, prefer `coinsPerUtxoSize`
+    * `coinsPerUtxoSize` is now
+      * Cost per UTxO **word** for Alonzo.
+      * Cost per UTxO **byte** for Babbage and later.
 
 # Version [0.5.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.4.0.1...v0.5.0.0) (2022-06-06)